requirements.txt
aiogram
python-dotenv
aiohttp
sqlalchemy
aiosqlite
pillow
requests
pandas
openpyxl
fastapi
uvicorn
starlette
yookassa
